Timing tweaks
